+++ deai-python/client/data.py
@@ -1,0 +1,30 @@
+import argparse
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+import torch.optim as optim
+from torchvision import datasets, transforms
+from torch.optim.lr_scheduler import StepLR
+import socket
+import json
+import time
+import zmq
+
+
+def load_data(num, total, train_kwargs, test_kwargs):
+    transform = transforms.Compose([
+        transforms.ToTensor(),
+        transforms.Normalize((0.1307,), (0.3081,))
+    ])
+
+    train_set = datasets.MNIST('./data', train=True, download=True,
+                               transform=transform)
+    test_set = datasets.MNIST('./data', train=False,
+                              transform=transform)
+
+    # Use a subset of the whole training set
+    indices = range(len(train_set) / total * num,
+                    len(train_set) / total * (num + 1))
+    train_set = torch.utils.data.Subset(train_set, indices)
+
+    return train_set, test_set
