+++ deai-python/client/training.py
@@ -7,31 +7,8 @@ from torchvision import datasets, transforms
 from torch.optim.lr_scheduler import StepLR
 import socket
 import json
-
-HOST = 'localhost'
-PORT = 50007
-
-
-class Net(nn.Module):
-    def __init__(self):
-        super(Net, self).__init__()
-        self.conv1 = nn.Conv2d(1, 32, 3, 1)
-        self.conv2 = nn.Conv2d(32, 64, 3, 1)
-        self.fc1 = nn.Linear(9216, 128)
-        self.fc2 = nn.Linear(128, 10)
-
-    def forward(self, x):
-        x = self.conv1(x)
-        x = F.relu(x)
-        x = self.conv2(x)
-        x = F.relu(x)
-        x = F.max_pool2d(x, 2)
-        x = torch.flatten(x, 1)
-        x = self.fc1(x)
-        x = F.relu(x)
-        x = self.fc2(x)
-        output = F.log_softmax(x, dim=1)
-        return output
+import time
+import zmq
 
 
 def train(args, model, device, train_loader, optimizer, epoch):
@@ -51,126 +28,17 @@ def train(args, model, device, train_loader, optimizer, epoch):
                 break
 
 
-def test(model, device, test_loader):
-    model.eval()
-    test_loss = 0
-    correct = 0
-    with torch.no_grad():
-        for data, target in test_loader:
-            data, target = data.to(device), target.to(device)
-            output = model(data)
-            # sum up batch loss
-            test_loss += F.nll_loss(output, target, reduction='sum').item()
-            # get the index of the max log-probability
-            pred = output.argmax(dim=1, keepdim=True)
-            correct += pred.eq(target.view_as(pred)).sum().item()
-
-    test_loss /= len(test_loader.dataset)
-
-    print('\nTest set: Average loss: {:.4f}, Accuracy: {}/{} ({:.0f}%)\n'.format(
-        test_loss, correct, len(test_loader.dataset),
-        100. * correct / len(test_loader.dataset)))
-
-
-def send_model_state(tcp_socket, model):
-    message = json.dumps(model.state_dict()).encode('utf-8')
-    tcp_socket.sendall(message)
-
-
-def average_model(model, received_state):
+def average_with_model(model, received_state):
     model_state = model.state_dict()
-    for key in received_state:
+    for key in model_state:
         model_state[key] = (model_state[key] + received_state[key]) / 2
     model.load_state_dict(model_state)
 
 
-def main():
-    # Training settings
-    parser = argparse.ArgumentParser(description='PyTorch MNIST Example')
-    parser.add_argument('--batch-size', type=int, default=64, metavar='N',
-                        help='input batch size for training (default: 64)')
-    parser.add_argument('--test-batch-size', type=int, default=1000, metavar='N',
-                        help='input batch size for testing (default: 1000)')
-    parser.add_argument('--epochs', type=int, default=14, metavar='N',
-                        help='number of epochs to train (default: 14)')
-    parser.add_argument('--lr', type=float, default=1.0, metavar='LR',
-                        help='learning rate (default: 1.0)')
-    parser.add_argument('--gamma', type=float, default=0.7, metavar='M',
-                        help='Learning rate step gamma (default: 0.7)')
-    parser.add_argument('--no-cuda', action='store_true', default=False,
-                        help='disables CUDA training')
-    parser.add_argument('--dry-run', action='store_true', default=False,
-                        help='quickly check a single pass')
-    parser.add_argument('--seed', type=int, default=1, metavar='S',
-                        help='random seed (default: 1)')
-    parser.add_argument('--log-interval', type=int, default=10, metavar='N',
-                        help='how many batches to wait before logging training status')
-    parser.add_argument('--save-model', action='store_true', default=False,
-                        help='For Saving the current Model')
-    # The number is used to decide which part of the original dataset is trained on this peer
-    parser.add_argument('--num', type=int, default=0, metavar='N',
-                        help='Number of this peer')
-    parser.add_argument('--total', type=int, default=1, metavar='N',
-                        help='Total number of peers')
-
-    args = parser.parse_args()
-    use_cuda = not args.no_cuda and torch.cuda.is_available()
-    torch.manual_seed(args.seed)
-    device = torch.device("cuda" if use_cuda else "cpu")
-    train_kwargs = {'batch_size': args.batch_size}
-    test_kwargs = {'batch_size': args.test_batch_size}
-    if use_cuda:
-        cuda_kwargs = {'num_workers': 1,
-                       'pin_memory': True,
-                       'shuffle': True}
-        train_kwargs.update(cuda_kwargs)
-        test_kwargs.update(cuda_kwargs)
-
-    num = args.num
-    total = args.total
-
-    transform = transforms.Compose([
-        transforms.ToTensor(),
-        transforms.Normalize((0.1307,), (0.3081,))
-    ])
-    train_set = datasets.MNIST('./data', train=True, download=True,
-                               transform=transform)
-    test_set = datasets.MNIST('./data', train=False,
-                              transform=transform)
-
-    # Use a subset of the whole training set
-    indices = range(len(train_set) / total * num,
-                    len(train_set) / total * (num + 1))
-    train_set = torch.utils.data.Subset(train_set, indices)
-
-    train_loader = torch.utils.data.DataLoader(train_set, **train_kwargs)
-    test_loader = torch.utils.data.DataLoader(test_set, **test_kwargs)
-
-    model = Net().to(device)
-    optimizer = optim.SGD(model.parameters(), lr=args.lr)
-    #scheduler = StepLR(optimizer, step_size=1, gamma=args.gamma)
-
-    tcp_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    tcp_socket.bind((HOST, PORT))
-    tcp_socket.listen(5)
-
-    for epoch in range(1, args.epochs + 1):
-        train(args, model, device, train_loader, optimizer, epoch)
-        # test(model, device, test_loader)
-        send_model_state(tcp_socket, model)
-        # scheduler.step()
-
-        while data = tcp_socket.recv(8192):
-            total_data = [data]
-            while data = tcp_socket.recv(8192):
-                total_data.append(data)
-            message = ''.join(total_data)
-            received_state = json.loads(message.decode('utf-8'))
-            average_model(model, received_state)
-
-    if args.save_model:
-        torch.save(model.state_dict(), "mnist_cnn.pt")
-
-
-if __name__ == '__main__':
-    main()
+def average_with_models(model, received_states):
+    total_num = len(received_states) + 1
+    model_state = model.state_dict()
+    for key in model_state:
+        model_state[key] = (model_state[key] + sum([state[key]
+                                                    for state in received_states])) / total_num
+    model.load_state_dict(model_state)
