+++ deai-python/server/main.py
@@ -1,0 +1,2 @@
+import register
+import zmq
